move context back to webui file
