+++ webui/webui.go
@@ -5,8 +5,10 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
+	"time"
 
 	"github.com/gocraft/web"
+	"github.com/gojek/work"
 	"github.com/gojek/work/webui/internal/assets"
 )
 
@@ -16,6 +18,138 @@ func mustAsset(name string) []byte {
 		panic(err)
 	}
 	return b
+}
+
+type context struct {
+	client *work.Client
+}
+
+func (c *context) ping(rw web.ResponseWriter, _ *web.Request) {
+	render(rw, map[string]string{"ping": "pong", "current_time": time.Now().Format(time.RFC3339)}, nil)
+}
+
+func (c *context) queues(rw web.ResponseWriter, _ *web.Request) {
+	response, err := c.client.Queues()
+	render(rw, response, err)
+}
+
+func (c *context) workerPools(rw web.ResponseWriter, _ *web.Request) {
+	response, err := c.client.WorkerPoolHeartbeats()
+	render(rw, response, err)
+}
+
+func (c *context) busyWorkers(rw web.ResponseWriter, _ *web.Request) {
+	observations, err := c.client.WorkerObservations()
+	if err != nil {
+		renderError(rw, err)
+		return
+	}
+
+	var busyObservations []*work.WorkerObservation
+	for _, ob := range observations {
+		if ob.IsBusy {
+			busyObservations = append(busyObservations, ob)
+		}
+	}
+
+	render(rw, busyObservations, err)
+}
+
+func (c *context) retryJobs(rw web.ResponseWriter, r *web.Request) {
+	page, err := parsePage(r)
+	if err != nil {
+		renderError(rw, err)
+		return
+	}
+
+	jobs, count, err := c.client.RetryJobs(page)
+	if err != nil {
+		renderError(rw, err)
+		return
+	}
+
+	response := struct {
+		Count int64            `json:"count"`
+		Jobs  []*work.RetryJob `json:"jobs"`
+	}{Count: count, Jobs: jobs}
+
+	render(rw, response, err)
+}
+
+func (c *context) scheduledJobs(rw web.ResponseWriter, r *web.Request) {
+	page, err := parsePage(r)
+	if err != nil {
+		renderError(rw, err)
+		return
+	}
+
+	jobs, count, err := c.client.ScheduledJobs(page)
+	if err != nil {
+		renderError(rw, err)
+		return
+	}
+
+	response := struct {
+		Count int64                `json:"count"`
+		Jobs  []*work.ScheduledJob `json:"jobs"`
+	}{Count: count, Jobs: jobs}
+
+	render(rw, response, err)
+}
+
+func (c *context) deadJobs(rw web.ResponseWriter, r *web.Request) {
+	page, err := parsePage(r)
+	if err != nil {
+		renderError(rw, err)
+		return
+	}
+
+	jobs, count, err := c.client.DeadJobs(page)
+	if err != nil {
+		renderError(rw, err)
+		return
+	}
+
+	response := struct {
+		Count int64           `json:"count"`
+		Jobs  []*work.DeadJob `json:"jobs"`
+	}{Count: count, Jobs: jobs}
+
+	render(rw, response, err)
+}
+
+func (c *context) deleteDeadJob(rw web.ResponseWriter, r *web.Request) {
+	diedAt, err := strconv.ParseInt(r.PathParams["died_at"], 10, 64)
+	if err != nil {
+		renderError(rw, err)
+		return
+	}
+
+	err = c.client.DeleteDeadJob(diedAt, r.PathParams["job_id"])
+
+	render(rw, map[string]string{"status": "ok"}, err)
+}
+
+func (c *context) retryDeadJob(rw web.ResponseWriter, r *web.Request) {
+	diedAt, err := strconv.ParseInt(r.PathParams["died_at"], 10, 64)
+	if err != nil {
+		renderError(rw, err)
+		return
+	}
+
+	err = c.client.RetryDeadJob(diedAt, r.PathParams["job_id"])
+
+	render(rw, map[string]string{"status": "ok"}, err)
+}
+
+func (c *context) deleteAllDeadJobs(rw web.ResponseWriter, r *web.Request) {
+	err := c.client.DeleteAllDeadJobs()
+	render(rw, map[string]string{"status": "ok"}, err)
+}
+
+func (c *context) retryAllDeadJobs(rw web.ResponseWriter, r *web.Request) {
+	err := c.client.RetryAllDeadJobs()
+	render(rw, map[string]string{"status": "ok"}, err)
 }
 
 func render(rw web.ResponseWriter, jsonable interface{}, err error) {
